+++ NNHybridRN/components/refresh/RefreshFlatList.js
@@ -7,7 +7,6 @@ import {
     FlatList,
     ActivityIndicator,
     Animated,
-    Easing
 } from 'react-native';
 import { PropTypes } from 'prop-types';
 import AppUtil from '../../utils/AppUtil';
@@ -17,31 +16,27 @@ import {
     FooterRefreshState,
     defaultHeaderProps,
     defaultFooterProps,
-    defaultHeaderRefreshComponent,
-    defaultFooterRefreshComponent
 } from './RefreshConst';
-import { TouchableOpacity } from 'react-native-gesture-handler';
-
-const arrowImage = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAABQBAMAAAD8TNiNAAAAJ1BMVEUAAACqqqplZWVnZ2doaGhqampoaGhpaWlnZ2dmZmZlZWVmZmZnZ2duD78kAAAADHRSTlMAA6CYqZOlnI+Kg/B86E+1AAAAhklEQVQ4y+2LvQ3CQAxGLSHEBSg8AAX0jECTnhFosgcjZKr8StE3VHz5EkeRMkF0rzk/P58k9rgOW78j+TE99OoeKpEbCvcPVDJ0OvsJ9bQs6Jxs26h5HCrlr9w8vi8zHphfmI0fcvO/ZXJG8wDzcvDFO2Y/AJj9ADE7gXmlxFMIyVpJ7DECzC9J2EC2ECAAAAAASUVORK5CYII=';
-
-
 
 /**
  * 默认刷新头部组件
  */
-const arrowOrActivityComponent = (headerRefreshState, arrowAnimation) => {
+const headerArrowOrActivity = (headerRefreshState, arrowAnimation) => {
     if (headerRefreshState == HeaderRefreshState.Refreshing) {
         return (
-            <ActivityIndicator style={{ marginRight: 10 }} />
+            <ActivityIndicator
+                style={{ marginRight: 10 }}
+                size="small"
+                color={AppUtil.app_theme}
+            />
         );
     } else {
         return (
             <Animated.Image
-                source={{ uri: arrowImage }}
-                resizeMode={'contain'}
+                source={require('../../resource/images/arrow/refresh_arrow.png')}
                 style={{
-                    width: 14,
-                    height: 23,
+                    width: 20,
+                    height: 20,
                     marginRight: 10,
                     transform: [{
                         rotateZ: arrowAnimation.interpolate({
@@ -84,6 +79,63 @@ const headerTitleComponent = (headerRefreshState, props) => {
             {headerTitle}
         </Text>
     );
+}
+
+// 默认加载更多组件
+export const defaultFooterRefreshComponent = ({
+    footerRefreshState,
+    footerRefreshingText,
+    footerFailureText,
+    footerNoMoreDataText,
+    footerEmptyDataText,
+    onHeaderRefresh,
+    onFooterRefresh,
+    data }) => {
+    switch (footerRefreshState) {
+        case FooterRefreshState.Idle:
+            return (
+                <View style={styles.footerContainer} />
+            );
+        case FooterRefreshState.Refreshing:
+            return (
+                <View style={styles.footerContainer} >
+                    <ActivityIndicator size="small" color={AppUtil.app_theme} />
+                    <Text style={[styles.footerText, { marginLeft: 7 }]}>
+                        {footerRefreshingText}
+                    </Text>
+                </View>
+            );
+        case FooterRefreshState.Failure:
+            return (
+                <TouchableOpacity onPress={() => {
+                    if (AppUtil.isEmptyArray(data)) {
+                        onHeaderRefresh && onHeaderRefresh();
+                    } else {
+                        onFooterRefresh && onFooterRefresh();
+                    } Î
+                }}>
+                    <View style={styles.footerContainer}>
+                        <Text style={styles.footerText}>{footerFailureText}</Text>
+                    </View>
+                </TouchableOpacity>
+            );
+        case FooterRefreshState.EmptyData:
+            return (
+                <TouchableOpacity onPress={() => { onHeaderRefresh && onHeaderRefresh(); }}>
+                    <View style={styles.footerContainer}>
+                        <Text style={styles.footerText}>{footerEmptyDataText}</Text>
+                    </View>
+                </TouchableOpacity>
+            );
+        case FooterRefreshState.NoMoreData:
+            return (
+                <View style={styles.footerContainer} >
+                    <Text style={styles.footerText}>{footerNoMoreDataText}</Text>
+                </View>
+            );
+    }
+
+    return null;
 }
 
 export default class RefreshFlatList extends Component {
@@ -144,28 +196,20 @@ export default class RefreshFlatList extends Component {
 
     }
 
-    shouldComponentUpdate(nextProps, nextState) {
-        if (nextProps === this.props && nextState === this.state) return false;
-        // if (nextProps.headerIsRefreshing === false && nextState.headerRefreshState === this.state.headerRefreshState) {
-        //     return false;
-        // }
-
-        return true;
-    }
-
     componentWillReceiveProps(nextProps) {
         const { headerIsRefreshing, listRef } = nextProps;
 
 
         if (headerIsRefreshing !== this.props.headerIsRefreshing) {
             // console.log('调用一下'+ headerIsRefreshing + this.props.headerIsRefreshing);
-
             const offset = headerIsRefreshing ? -this.headerHeight : 0;
             const headerRefreshState = headerIsRefreshing ? HeaderRefreshState.Refreshing : HeaderRefreshState.Idle;
 
+            if (!headerIsRefreshing) this.state.arrowAnimation.setValue(0);
+
             this.refs[listRef].scrollToOffset({ animated: true, offset });
             this.setState({ headerRefreshState });
-        } 
+        }
     }
 
     /**
@@ -190,7 +234,7 @@ export default class RefreshFlatList extends Component {
                     marginTop: -this.headerHeight,
                     height: this.headerHeight
                 }} >
-                    {arrowOrActivityComponent(headerRefreshState, arrowAnimation)}
+                    {headerArrowOrActivity(headerRefreshState, arrowAnimation)}
                     {headerTitleComponent(headerRefreshState, this.props)}
                 </View >
             );
@@ -203,17 +247,7 @@ export default class RefreshFlatList extends Component {
     _renderFooter = () => {
         const {
             footerRefreshState,
-
-            // footerRefreshingText,
-            // footerFailureText,
-            // footerNoMoreDataText,
-            // footerEmptyDataText,
-
             footerRefreshComponent,
-
-            // onHeaderRefresh,
-            // onFooterRefresh,
-            // data,
         } = this.props;
 
         if (footerRefreshComponent) {
@@ -222,60 +256,6 @@ export default class RefreshFlatList extends Component {
         }
 
         return defaultFooterRefreshComponent({ ...this.props });
-        // switch (footerRefreshState) {
-        //     case FooterRefreshState.Idle: {
-        //         return (
-        //             <View style={styles.footerContainer} />
-        //         );
-        //     }
-        //     case FooterRefreshState.Refreshing: {
-        //         return (
-        //             <View style={styles.footerContainer} >
-        //                 <ActivityIndicator size="small" color={AppUtil.app_theme} />
-        //                 <Text style={[styles.footerText, { marginLeft: 7 }]}>
-        //                     {footerRefreshingText}
-        //                 </Text>
-        //             </View>
-        //         );
-        //     }
-        //     case FooterRefreshState.Failure: {
-        //         const pressHandler = () => {
-        //             if (AppUtil.isEmptyArray(data)) {
-        //                 onHeaderRefresh && onHeaderRefresh();
-        //             } else {
-        //                 onFooterRefresh && onFooterRefresh();
-        //             }
-        //         };
-
-        //         return (
-        //             <TouchableOpacity onPress={() => pressHandler()}>
-        //                 <View style={styles.footerContainer}>
-        //                     <Text style={styles.footerText}>{footerFailureText}</Text>
-        //                 </View>
-        //             </TouchableOpacity>
-        //         );
-        //     }
-        //     case FooterRefreshState.EmptyData: {
-        //         return (
-        //             <TouchableOpacity onPress={() => { onHeaderRefresh && onHeaderRefresh(); }}>
-        //                 <View style={styles.footerContainer}>
-        //                     <Text style={styles.footerText}>{footerEmptyDataText}</Text>
-        //                 </View>
-        //             </TouchableOpacity>
-        //         );
-        //     }
-        //     case FooterRefreshState.NoMoreData: {
-        //         return (
-        //             <View style={styles.footerContainer} >
-        //                 <Text style={styles.footerText}>
-        //                     {footerNoMoreDataText}
-        //                 </Text>
-        //             </View>
-        //         );
-        //     }
-        // }
-
-        // return null;
     }
 
     render() {
@@ -283,7 +263,6 @@ export default class RefreshFlatList extends Component {
             <FlatList
                 {...this.props}
                 ref={this.props.listRef}
-                // scrollEventThrottle={16}
                 onScroll={event => this._onScroll(event)}
                 onMomentumScrollEnd={event => this._onMomentumScrollEnd(event)}
                 onScrollEndDrag={event => this._onScrollEndDrag(event)}
@@ -296,20 +275,6 @@ export default class RefreshFlatList extends Component {
         );
     }
 
-    // /**
-    //  * 手动调用下拉刷新
-    //  */
-    // beginRefresh() {
-    //     this._headerBeginRefresh();
-    // }
-
-    // /**
-    //  * 手动结束
-    //  */
-    // endRefresh() {
-    //     this._headerEndRefresh();
-    // }
-
     /**
      * 列表正在滚动
      * @private
@@ -317,27 +282,16 @@ export default class RefreshFlatList extends Component {
      */
     _onScroll(event) {
         this.offsetY = event.nativeEvent.contentOffset.y;
-        console.log('_onScroll');
-        console.log(this.state);
-        console.log(this.offsetY);
         if (this.isDragging) {
             if (!this._isRefreshing()) {
                 if (this.offsetY <= -this.headerHeight) {
                     // 松开以刷新
                     this.setState({ headerRefreshState: HeaderRefreshState.Pulling });
-                    // Animated.timing(this.state.arrowAnimation, {
-                    //     toValue: 1,
-                    //     duration: 50,
-                    //     easing: Easing.inOut(Easing.quad)
-                    // }).start();
+                    this.state.arrowAnimation.setValue(1);
                 } else {
                     // 下拉以刷新
                     this.setState({ headerRefreshState: HeaderRefreshState.Idle });
-                    // Animated.timing(this.state.arrowAnimation, {
-                    //     toValue: 0,
-                    //     duration: 100,
-                    //     easing: Easing.inOut(Easing.quad)
-                    // }).start();
+                    this.state.arrowAnimation.setValue(0);
                 }
             }
         }
@@ -351,19 +305,16 @@ export default class RefreshFlatList extends Component {
     _onScrollBeginDrag(event) {
         this.isDragging = true;
         this.offsetY = event.nativeEvent.contentOffset.y;
-        console.log('_onScrollBeginDrag');
-        console.log(this.state);
-        console.log(this.offsetY);
     }
 
+    /**
+     * 列表结束拖拽
+     * @private
+     * @param {{}} event
+     */
     _onScrollEndDrag(event) {
         this.isDragging = false;
         this.offsetY = event.nativeEvent.contentOffset.y;
-        console.log('_onScrollEndDrag');
-        console.log(this.state);
-        console.log(this.offsetY);
-        
-
         const { listRef, onHeaderRefresh } = this.props;
 
         if (!this._isRefreshing()) {
@@ -376,37 +327,16 @@ export default class RefreshFlatList extends Component {
             if (this.offsetY <= 0) {
                 this.refs[listRef].scrollToOffset({ animated: true, offset: -this.headerHeight });
             }
-        }   
+        }
     }
+
     /**
-     * 列表结束拖拽
+     * 列表停止滚动
      * @private
      * @param {{}} event
      */
     _onMomentumScrollEnd(event) {
         this.offsetY = event.nativeEvent.contentOffset.y;
-        console.log('_onMomentumScrollEnd');
-        console.log(this.state);
-        console.log(this.offsetY);
-        // this.isDragging = false;
-        // this.offsetY = event.nativeEvent.contentOffset.y;
-        // console.log(this.state);
-        // console.log(this.offsetY);
-        
-
-        // const { listRef, onHeaderRefresh } = this.props;
-
-        // if (!this._isRefreshing()) {
-        //     if (this.state.headerRefreshState === HeaderRefreshState.Pulling) {
-        //         this.refs[listRef].scrollToOffset({ animated: true, offset: -this.headerHeight });
-        //         this.setState({ headerRefreshState: HeaderRefreshState.Refreshing });
-        //         onHeaderRefresh && onHeaderRefresh();
-        //     }
-        // } else {
-        //     if (this.offsetY <= 0) {
-        //         this.refs[listRef].scrollToOffset({ animated: true, offset: -this.headerHeight });
-        //     }
-        // }
     }
 
     /**
@@ -418,37 +348,6 @@ export default class RefreshFlatList extends Component {
             this.props.footerRefreshState === FooterRefreshState.Refreshing
         );
     }
-
-    /**
-     * 头部刷新结束
-     */
-    _headerEndRefresh() {
-        this.setState({ headerRefreshState: HeaderRefreshState.Idle });
-
-        // Animated.timing(this.state.arrowAnimation, {
-        //     toValue: 0,
-        //     duration: 100,
-        //     easing: Easing.inOut(Easing.quad)
-        // }).start();
-
-        const { listRef } = this.props;
-        this.refs[listRef].scrollToOffset({ animated: true, offset: 0 });
-    }
-
-    // /**
-    //  * 触发头部刷新
-    //  */
-    // _headerBeginRefresh = () => {
-    //     const { onHeaderRefresh, listRef } = this.props;
-
-    //     if (this._isRefreshing()) return;
-
-    //     this.refs[listRef].scrollToOffset({ animated: true, offset: -this.headerHeight });
-
-    //     this.setState({ headerRefreshState: HeaderRefreshState.Refreshing });
-
-    //     onHeaderRefresh && onHeaderRefresh();
-    // }
 
     /**
      * 触发加载更多
